fix(build): fix -lutil check on FreeBSD
On Linux, we need to add -lutil to the linker flags. To do this, we
try-compile a small program which calls forkpty().

On FreeBSD, we also need to add -lutil to the linker flags. Our
try-compile script fails to build on FreeBSD because it includes
<pty.h>, which is Linux-only. This leads to a build failure:

    ld: error: undefined symbol: forkpty
    >>> referenced by test-cli.cpp
    >>>               [snip]/test-cli.cpp.o:(quick_lint_js::(anonymous namespace)::test_cli_auto_hyperlinks_are_not_printed_in_dumb_terminals_Test::TestBody())

Tweak the try-compile to include whatever headers are necessary for
forkpty() depending on the platform:

* <libutil.h> on FreeBSD
* <pty.h> on Linux
* <util.h> on macOS

(-lutil is not needed on macOS, but I included the <util.h> for
completeness.)

This patch fixes the FreeBSD build.

Co-Authored-By: Nico Sonack <nsonack@herrhotzenplotz.de>