fix(fe): treat 'x! == y' as an error, not warning, in JS mode
'x! == y' is legal in TypeScript, but not in JavaScript. We suppress
E0261 ("TypeScript non-null assertions are not allowed in JavaScript")
in favor of E0373 ("unexpected whitespace between '!' and '=='") in this
example, but E0373 is a warning, not an error.

Create a new diagnostic, E0374, which is like E0373 but is an error
instead of a warning. Also, remove references to TypeScript in E0374's
documentation.