refactor(ci): split VS Code build jobs from test jobs
Currently, we build the VS Code extension then immediately test it
(except on Linux-x86_64). This is problematic because the .vsix might be
missing some files (such as MinGW DLLs) if we build incorrectly.

Test the VS Code extension in a separate job. This makes every platform
test similar to Linux-x86_64 before this patch.