fix: fix build failure with recent MinGW GCC
Building quick-lint-js-test.exe with MinGW GCC 13.1.0 (Windows x64) LTO fails
with a multiple definition error:

    ld.exe: linter.cpp.obj (symbol from plugin): multiple definition of `bool quick_lint_js::try_catch_stack<quick_lint_js::parser::fatal_parse_error>::try_catch<bool, quick_lint_js::parser::catch_fatal_parse_errors<[snip]';
    test-parse-typescript-class.cpp.obj: first defined here

This error is bogus because the function in question is implicitly
declared inline (COMDAT), so multiple definitions shouldn't be a
problem.

Work around the compiler bug by moving the definition of the function
into a .cpp file.