fix(vim): fix tracing logging too late with nvim-lspconfig
nvim-lspconfig takes a while to tell quick-lint-js about the user's
configuration. This means that LSP traffic might not get traced
according to the quick-lint-js.trace-directory setting until after
'textDocument/didOpen' notifications. This means that traces are usually
incomplete.

Teach quick-lint-js's LSP server to accept settings in the
'initialize' message. Use this new feature in the nvim-lspconfig
plugin. This causes quick-lint-js to promptly trace the 'initialized'
message and the first 'textDocument/didOpen' message.