fix(emacs): fix Flycheck commenting line not updating diagnostics
With the quick-lint-js Flycheck plugin, if you use `C-x C-;` to comment
a line, quick-lint-js does not run. This is because (strangely) Flycheck
considers commenting a line to be a 'new-line' event, not an
'idle-change' event (which most changes emit).

Make quick-lint-js run when commenting a line by enabling it for the
'new-line' event. Unfortunately, this is a documentation fix; existing
users will need to update their .emacs files.