refactor(fe): move visit_variable_declaration call for namespaces
Currently, parse_and_visit_typescript_namespace_head calls
visit_variable_declaration as soon as the namespace's name is found.
This makes it difficult to change variable_declaration_flags depending
on the contents of the namespace (which I will need to do in a future
patch to implement some diagnostics).

Factor the visit_variable_declaration call outside of
parse_and_visit_typescript_namespace_head so the flags can be better
controlled by the caller.