refactor(fe): declare namespace name after parsing body
In order to implement certain diagnostics properly, the parser needs to
communicate whether the namespace is empty or not. I want to do this
through variable_declaration_flags, but we cannot change flags after
parsing the body.

Change the parser to call visit_variable_declaration after
visit_exit_namespace_scope so that the variable_declaration_flags can
depend on features of the namespace's body.