refactor(diag): make diagnostic struct code more readable using code gen
Currently, diagnostic types are written using an ugly macro-based DSL.
This is ugly and sometimes inflexible.

Change the source code for diagnostic types to be mostly-readable plain
C++ code. Use C++ annotations, which are ignored by compilers, for
metadata such as the diagnostic code and messages. Write a script to
extract the C++ annotations and generate the old macro-based DSL.

Future changes will chip away at existing uses of QLJS_DIAG_TYPES and
eventually remove it.