refactor(diag): work around GCC 8.3 bug
GCC 8.3 and older (and also GCC 9.2 and older) have a bug where they
report "abstract_keyword was not declared in this scope" (for example).
Identifiers in C++ attributes should not be variable references, but
these buggy GCC versions treat them as references, failing the build.

Currently, this bug does not manifest because diagnostic-types-2.h is
not #include-d in any translation units. However, I want it to be
#include-d, so we need to make it compile with the buggy GCC.

(GCC 8.3 ships with Debian 10 Buster, which is our minimum supported
Debian version.)

(The GCC bug has been fixed in 8.4 and 9.3 and 10.1.)

Work around the GCC bug with two solutions:

* Scope severity by including 'Diagnostic_Severity::'.
* Write member variable references as string literals rather than
  identifiers. Use a macro to make this less silly-looking.