fix(i18n): prevent test failures when changing translatable string
If a locale (e.g. po/fr.po) contains an untranslated string not present
in the untranslated locale (messages.pot), compile-translations
generates a test which tries to translate that string. This test fails
because the table does not have a valid string for the untranslated
locale.

Fix this bug by using the list of all untranslated strings, ignoring the
untranslated locale.