refactor(container): switch Fixed_Vector back to a class (from union)
I want to make Fixed_Vector's destructor conditionally trivial, but this
is is only possible with a limited number of techniques:

1. Add a template parameter defaulted to
   std::is_trivially_destructible_v<T>. Specialize the class for true and
   false, copy-pasting the entire implementation (except the destructors
   will differ).
2. Do option 1, but with a base class.
3. Add a requires clause to the trivial destructor:

       ~Fixed_Vector() requires std::is_trivially_destructible_v<T> = default;

Option 1 is a maintenance headache. Option 2 is not possible because
Fixed_Vector is a union and a union cannot derive from a class or union.
Option 3 doesn't work due to a bug in Clang 14 and older.

Make option 2 possible by refactoring Fixed_Vector back into a class (as
it was originally).