refactor(debian): create 'all' Packages file for quick-lint-js-vim
binary-amd64/Packages and binary-arm64/Packages both have entries for
the quick-lint-js-vim package. I want there to be one listing of the
quick-lint-js-vim package for either architecture in case you write
[arch=amd64,arm64] in your sources.list. Move the quick-lint-js-vim
package into the binary-all/Packages file.

With this patch, because there are two bit-differing copies of
quick-lint-js-vim for 2.14.0 and 2.15.0, 'apt install quick-lint-js'
complains:

> W: Sources disagree on hashes for supposely identical version
> '2.15.0-1' of 'quick-lint-js-vim:amd64'.

This warning was already issued if your sources.list included
[arch=amd64,arm64]. Now, the warning is issued even if your sources.list
includes only [arch=amd64] or [arch=arm64]. This issue will be fixed in
a future commit.

This is a step toward fixing
https://github.com/quick-lint/quick-lint-js/issues/1048.