refactor(io): allow stopping event loop prematurely
Currently, Event_Loop only stops if the reader reaches end-of-file. This
is hard to do when benchmarking the Rome LSP server because it doesn't
seem to ever send EOF.

Teach Event_Loop to optionally shut down if requested by the derived
class. Currently this feature is unused, but it will be used soon in
benchmark-lsp-servers.