fix(benchmark-lsp): fix benchmarks hanging with Rome
On Linux, on the first run of benchmark-lsp-servers with Rome (i.e. when
Rome's daemon has not started yet), shutdown does not happen properly.
This is because Rome's daemon holds onto the pipes even after sending an
exit LSP notification.

Make the benchmark event loop exit early by disabling the reader when we
want to exit. This hack also works with Flow, so use it for Flow instead
of the existing close-the-writer hack.