refactor(io): implement a new event loop abstraction
Event_Loop is crummy. The CRTP pattern is gross, the interface for
registering file descriptors and handles is gross, it's very inflexible,
its semantics poorly documented, and testing is weak. Event_Loop was
designed to be lean, but after adding several features to it for
different use cases, its design has proven to be bad.

Implement a new Event_Loop, called Event_Loop2. It is callback-based and
has explicit functions for registering events, disabling/enabling
events, and stopping the event loop. It is better documented (which is
easy because semantics are saner) and is better tested (because it's
easier to test the saner interface).

To prove the design can be ported, implement the same backends that
Event_Loop currently has:

* Windows: IOCP + threads
* macOS/FreeBSD: kqueue
* Linux: poll

Users of Event_Loop will gradually migrate to Event_Loop2 in future
commits.