fix(benchmark-lsp): avoid wasting CPU during benchmarks
Pipe write events are always enabled, so the event loop will keep
calling the on_pipe_write_ready callback unnecessarily on Linux and
macOS, wasting CPU and stealing resources from the LSP server we are
benchmarking. Fix this in the same way we avoid this problem with the
LSP server.