refactor(build): optimize incremental builds with write_file_if_different
If src/quick-lint-js/assert.cpp is modified, dependencies of
quick-lint-js-tool-lib, including
quick-lint-js-generate-diagnostic-metadata, are recompiled. This causes
quick-lint-js-generate-diagnostic-metadata to re-run during the build,
and quick-lint-js-generate-diagnostic-metadata recreates
quick-lint-js/diag/diagnostic-metadata-generated.h. This causes tons of
other things which #include this header to rebuild.

This rebuilding is overkill. Changing assert.cpp won't change the output
of diagnostic-metadata-generated.h.

Make quick-lint-js-generate-diagnostic-metadata not modify
diagnostic-metadata-generated.h if it doesn't need to be modified. This
signals to the Ninja build tool that its build graph needs to be
updated, causing Ninja to not rebuild #include-ers of
diagnostic-metadata-generated.h.