fix(io): make write_file_if_different not fail if file is missing
write_file_if_different should behave like write_file except it
shouldn't change mtimes if the contents match. If the file doesn't
exist, it should behave like write_file, but currently it fails with a
file-not-found error.

Fix write_file_if_different to treat a missing file the same as a file
with the wrong content.