fix(reflection): fix wrong location for diag codes
When I implemented string decoding, that put parsed string bodies on the
heap. If there are duplicate diagnostic codes, we give error_at a
pointer to a heap-allocated string which is not inside the source code.
Doing this leads to out-of-bounds source location information, so error
messages point nowhere interesting:

    diagnostic-types-2.h:3089:40903: error: diag code E0397 already in use; try this unused diag code: E0398
    diagnostic-types-2.h:3089:26935: note: Diag_TypeScript_Abstract_Static_Property used code E0397 here

Instead of using the error code string for location information, track
the location information explicitly. This fixes the error message's
locations:

    diagnostic-types-2.h:2826:23: error: diag code E0397 already in use; try this unused diag code: E0398
    diagnostic-types-2.h:1880:23: note: Diag_TypeScript_Abstract_Static_Property used code E0397 here