refactor(util): avoid repeated uri_base_name calls
When classifying URIs to see if they are .d.ts files or .tsx files, we
effectively call uri_base_name multiple times. This is ugly.

Combine classification into one step. This is a nicer interface in my
opinion and it avoids the duplicate uri_base_name issue.

Unfortunately, this change adds another inefficiency: we search for
.tsx even when it's unnecessary. I think this inefficiency is worth
the improved interface.