refactor(cli): move get_language out of File_To_Lint
get_language is a bit weird. There are two functions: a top-level
function and a member function of File_To_Lint. Make both versions
top-level functions. This makes it more obvious that the functions do
the same thing, and will make it more natural to add some features to
get_language later.