fix(emacs): clear diagnostics properly in Flymake
When our Flymake code set diagnostics, it set them for the whole
buffer with :range. However, the :range was incorrect; the begin/end
points given were for the *stdout* temporary buffer, not for the
user's code buffer. This meant that Flymake diagnostics markers were
often not removed because they were not inside the range. (stdout is
often much smaller than the JavaScript source being linted.)

Fix this by using :range with points derived from the correct buffer.