refactor(container): put vector profiler code into non-debug
All the vector profiling code lives in quick-lint-js-lib-debug.
quick-lint-js-lib-debug depends on quick-lint-js-lib. This is awkward
because it means that, if vector profiling is enabled,
quick-lint-js-lib depends on quick-lint-js-lib-debug. This is a
dependency cycle, but for some reason CMake doesn't complain.

Break the dependency cycle by putting the main vector profiler code
into quick-lint-js-lib, keeping the <iostream>-using code (used by
tests only) in quick-lint-js-lib-debug.