feat(website): improve usability of LSP docs
I suspect people are visiting the LSP documentation page on the
website expecting to see instructions on how to install the
quick-lint-js LSP plugins. Our documentation for these tasks is in the
install guides.

Add links to the install guides at the top of the LSP documentation
page to guide the user to the right spot.