fix(typescript): fix variable conflict check confused by type
If a type and a run-time variable have the same name, and if either is
redeclared, quick-lint-js only checks the first declaration when
checking for conflicting declarations. This leads to false negatives
in cases such as the following:

    type X = null;
    let X;
    let X;  // This checks 'type X' but not 'let X'

Fix this by checking all prior declarations rather than just the first.