refactor(fe): change ambiguity resolution of '(' in types
Currently, when we see '(' in a type, we parse it as a parenthesized
type. If we then see '=>', or if parsing as a parenthesized type
failed, we re-parse it as an arrow type.

Do it the other way (parse as an arrow type first, falling back to a
parenthesized type on error) for a few reasons:

* After some experimentation, this seems to be what TypeScript does.
  This might affect some edge cases.
* This unlocks some future cleanup regarding allow_parenthesized_type.
* Performance might be better (unmeasured) because arrow types are
  more common than parenthesized types (unmeasured).