refactor(fe): split code paths for ':' parsing in expressions
When ':' is encountered in an expression, there are three major code
paths:

* Do nothing (stop parsing)
* Parse a type annotation
* Parse a type annotation if it's an arrow function's return type
  annotation (used for disambiguating '? :' expressions)

These code paths have similarities but also important differences. I
want to add more differences, but that is difficult with the current
shared code structure. Refactor the code to duplicate the similarities
so that the code paths can diverge independently.