fix(typescript): fix false positive with import and var with same name
The following is valid TypeScript if the imported x is a type-only
declaration (e.g. an interface):

    import {x} from './mod';
    x = 42;
    var x;

quick-lint-js incorrectly emits a warning for this code. Fix it.

Unfortunately, this fix introduces another bug where we do not
diagnose the following code:

    import {x} from './mod';
    x = 42;

I prefer the new bug because it does not reject valid code.