fix(typescript): parse 'aria-label' as single token after generic component
If a generic JSX component tag name, such as 'Foo<Bar>' in
'<Foo<Bar> />', is followed by a hyphenated attribute name, the hyphen
is not included in the token. In other words, tokenization is broken:

    <Foo<Bar> aria-label="baz" />
    ^                less
     ^^^             identifier
        ^            less
         ^^^         identifier
            ^        greater
              ^^^^   identifier
                  ^  minus

Fix this by calling skip_in_jsx() instead of skip() when skipping the
'>' in 'Foo<Bar>'.