refactor(test): require '{' and '}' to be escaped in assertions
I want to allow diagnostic assertions to contain multiple string
members. For example:

     u8"Diag_Foo{.bar=asdf}{.baz=qwer}"_diag

This isn't possible right now because we parse .bar's value as
"asdf}{.baz=qwer".

Change parsing to parse only until the first '}'. Allow '}' inside the
value by requiring it to be escaped. For consistency, also require '{'
to be escaped.